rag_module/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
huggingface-hub
sentence-transformers
scikit-learn
numpy
torch
transformers
requests
boto3
slowapi
python-dotenv
